--- src/Blog.js
@@ -1,15 +1,0 @@
-// A Link to a blog
-import React, { Component } from 'react';
-import './App.css';
-
-// const LocalWebURL = require('./blog/index.php');
-
-class Blog extends Component {
-    render() {
-        return (
-            <div>Hello</div>
-        );
-    }
-}
-
-export default Blog;
