+++ src/index.css
@@ -6,3 +6,32 @@ body {
 ol, ul {
   padding-left: 30px;
 }
+
+.Header {
+  text-align: center;
+}
+
+.Header-logo {
+  animation: Header-logo-spin infinite 20s linear;
+  height: 80px;
+}
+
+.Header-header {
+  background-color: #122;
+  height: 150px;
+  padding: 20px;
+  color: white;
+}
+
+.Header-title {
+  font-size: 1.5em;
+}
+
+.Header-intro {
+  font-size: large;
+}
+
+@keyframes Header-logo-spin {
+  from { transform: rotate(0deg); }
+  to { transform: rotate(359deg); }
+}
